// apps/web/src/api/generated/model/createProviderConnectionDtoType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Open WebUI Klon API
 * OpenAPI spec version: 0.0.0
 */

export type CreateProviderConnectionDtoType = typeof CreateProviderConnectionDtoType[keyof typeof CreateProviderConnectionDtoType];


export const CreateProviderConnectionDtoType = {
  ollama: 'ollama',
  openai_compatible: 'openai_compatible',
} as const;